+++ lib/generators/happy_seed/base/base_generator.rb
@@ -7,7 +7,7 @@ module HappySeed
         puts "Installing happy_seed:base environment"
         gem 'dotenv-rails', :groups=>[:development, :test]
         gem 'rdiscount', :groups => [:development, :test]
-        gem 'unicorn'
+        gem 'puma'
         gem 'rails_12factor'
 
         Bundler.with_clean_env do
@@ -22,21 +22,13 @@ module HappySeed
         inject_into_file 'config/environments/test.rb', "  config.log_level = :error\n", before: "end\n"
 
         begin
-          inject_into_file 'spec/spec_helper.rb', "\n  config.include Warden::Test::Helpers, type: :feature\n  config.include ControllerHelpers, type: :controller\n  Warden.test_mode!\n", :before => "\nend\n"
-          prepend_to_file 'spec/spec_helper.rb', "require_relative 'support/controller_helpers'\n"
-        rescue
-          say_status :spec, "Unable to add login helpers to spec_helper.rb"
-        end
-
-
-        begin
           inject_into_file 'spec/rails_helper.rb', "require 'webmock/rspec'\n", after: "'rspec/rails'\n"
         rescue
           say_status :spec, "Unable to add webmock to rails_helper.rb", :red
         end
 
         begin
-          inject_into_file 'spec/rails_helper.rb', "\n  config.include Devise::TestHelpers, type: :controller\n  config.include Warden::Test::Helpers, type: :feature\n  config.include FactoryGirl::Syntax::Methods\n", :before => "\nend\n"
+          inject_into_file 'spec/rails_helper.rb', "\n  config.include FactoryGirl::Syntax::Methods\n", :before => "\nend\n"
           append_to_file 'spec/rails_helper.rb', "\nVCR.configure do |c|\n  c.cassette_library_dir  = Rails.root.join('spec', 'vcr')\n  c.hook_into :webmock\nend\n"
         rescue
           say_status :spec, "Unable to add factory girl and VCR to rails_helper.rb", :red