+++ lib/generators/happy_seed/devise/devise_generator.rb
@@ -32,11 +32,23 @@ module HappySeed
         
         begin
           prepend_to_file 'spec/spec_helper.rb', "require 'devise'\n"
+          prepend_to_file 'spec/spec_helper.rb', "require_relative 'support/controller_helpers'\n"
           inject_into_file 'spec/spec_helper.rb', "\n  config.include Devise::TestHelpers, type: :controller\n", :before => "\nend\n"
-
+          inject_into_file 'spec/spec_helper.rb', "\n  config.include Warden::Test::Helpers, type: :feature\n  config.include ControllerHelpers, type: :controller\n  Warden.test_mode!\n", :before => "\nend\n"
         rescue
           say_status :spec, "Unable to add devise helpers to spec_helper.rb", :red
         end
+
+        begin
+          inject_into_file 'spec/rails_helper.rb', "\n  config.include Devise::TestHelpers, type: :controller\n  config.include Warden::Test::Helpers, type: :feature\n", :before => "\nend\n"
+        rescue
+          say_status :spec, "Unable to add devise helpers to rails_helper.rb", :red
+        end
+
+        append_to_file "features/support/env.rb", "
+Warden.test_mode! 
+World(Warden::Test::Helpers)
+After{ Warden.test_reset! }"
 
         directory 'app'
         directory 'docs'