+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 Version 0.0.17
+  - Moved from unicorn to puma
   - Cleaned up twitter generator
   - Changed angular generator to move @app into the global scope
   - Changed angular generator to define controllers via @app.controller
+  - Moved Devise testing support from base_generator to devise_generator
+  - Moved happy_seed to development section of Gemfile
   
 Version 0.0.16
   - Fixes for rbenv (require bundler)